+++ server/src/game_server/objects/human.ts
@@ -380,6 +380,7 @@ export class Human extends MovingBody{
         if(preset.modifiers)this.temp_modifiers=preset.modifiers
         if(preset.inventory)this.inventory.load_preset(preset.inventory)
         if(preset.position)this.position=preset.position
+        if(preset.group_color)this.team_data.color=preset.group_color
 
         this.update_modifiers()